added minor interface codes and edits

+++ oxdb-engine/src/config/default.rs
@@ -1,3 +1,3 @@
 // Constants
-pub const BLOCK_SIZE: usize = 4096; // 4KB fallback block size
+pub const BLOCK_SIZE: u16 = 4096; // 4KB fallback block size
 pub const MAX_PAGE_LEN: u16 = 16; // 64 KB
